cleaned up code
Most of the codes were not well grouped, making it hard to read.
In this version, some of the previous lines have been grouped to
form functions. Some function names have changed as well to make it
more readable. All others are grouped well to enhance the reading of
the code.

There are few variables that I am not sure what they do yet... but
I am still going through.